Update Project 1 a
This experiment is to adjust the on-off and flashing rates of the LED on the STM32F103C8T6 board using 2 buttons. Initially, the LED in pinC13 is in the off state, and when the 1st button is pressed, the LED blinks in a 2000ms period. If the 1st button is pressed again during the blinking action, it turns the LED off. If the 2nd button is pressed while the LED is blinking with a 2000ms period, the period is halved (1000ms = 500ms LED on + 500 ms LED off). If the 2nd button is pressed while the LED is in the 1000ms period, the period returns to its previous state (2000ms).