Rename Project 2 to Project 1 b
STM32F103C8T6 board, one button three external LEDs (red, yellow, and green), a 10k ohm pull-down resistor, and 100-ohm resistors per led (3) were used in this experiment. Finite state machine application was designed with these components. Switch-case algorithm was used for the finite state machine. 10k ohm resistor was used for the button and 100-ohm resistors were used for LEDs. In the rest when the button is pressed, all three LEDs are turned on until releasing the button. When the button is released, LEDs are blinking respectively red, yellow, and green. When the button is again pressed, all LEDs are blinking at the same time. The blinking period was set for 1 second. If the button is released, all LEDs are turned off and the finite state machine was completed.