Rename Project 3 to Project 2 a
Aim: Finite state machine design and coding for traffic light control application using external
interrupts.
In the first part of the experiment, we are implementing a basic traffic light design using
STM32F103C8T6 board. In our circuit, we have 3 leds, and resistors for each of them,
respcetively. We also have a button to control our traffic light. When there’s no interrupt coming
from the button, the traffic light stays red for 5 seconds, followed by 2 seconds yellow, 5 seconds
green and 2 seconds yellow and so on.. If a pedestrian interrupts the traffic light cycle by pressing
the button while the light is green, the light immediately jumps to yellow and continues its
regular cycle. The system ignores the interrupt if the pedestrian presses the button when the light
is yellow or red.