Create Project 4
In this experiment, our main focus is to create a project that measures temperature via LM75A temperature sensor, and sends the related temperature data to the terminal program, for which purpose we have used Arduino Serial Port Terminal, via USART. Additionally, we have implemented a potentiometer, so that the user can set a threshold between 0 to 50 degrees. In order to set the threshold, the button needs to be pressed by the user beforehand. The data from the temperature sensor is sent to the terminal program through USART with 1 second intervals. The resolution of the temperature values is setted to be 0.5 degree celcius.

In addition to this, we also have an LED which turns on if the current temperature value is higher than the selected threshold and a 1 is sent from the terminal program, and turns off if it is less than the selected threshold or a 0 is sent from the terminal program.