Create Midterm Project a
The main objective in TASK I is to create a fundamental LED flashlight with a single button to control the dynamics of the LED. When the LED is in off state, a single short click to the button turns the LED on with the last adjusted brightness, 20% of brightness initially. From this point on, a single short click to the button closes the LED, a long press will aim to adjust the brightness to the desired level, first long press increases brightness, second long press decreases brightness and so forth. A double click will start the blinking dynamic of the LED with 500 ms of period