Create Final Project
The main task in this final experiment is to create a temperature controller that mimics a real-world industrial temperature controller. Our circuit consists of a temperature sensor, 2 buttons in order to adjust the desired temperature, a green LED which represents the cooler, a red LED which represents the heater and a yellow LED which represents the saturation point. Additionally, we use the USB to TTL UART converter which provides the user the information about the current condition of the system such as whether the system is currently heating or cooling, the current PWM duty ratio of the system, and current temperature as it is measured from the LM75A temperature sensor which is connected over I2C bus. We continuously measure the current temperature of the environment via our temperature sensor, calculate the error by taking the difference between the desired temperature, which is adjusted by the user using the buttons (increasing by +0.5 degrees when button 1 is released and decreasing by 0.5 degrees when button 2 is released), at which point we boost the error by our proportional controller with gain = 5. We then check the error value and take the necessary actions with our heater and cooler. We used systick timer to accurately measure the time elapsed in our program. In every 500 ms, we increase a flag counter called time_500ms_counter inside our Systick_Handler function. This way, we can adjust the sampling time of our program, where we let the user know about the current status of our system like whether it is currently operating at cooling mode or heating mode, current PWM duty ratio and also the current measured temperature.